Numerous improvements and bugfixes.
+ Role of the wrangler was expanded to be a clean provider of all data.

- Code concerned with data de-mangling was moved into the wranger.

* Renamed some variables to be more clear about purpose, and separated
the averaging step from the projection step more clearly.

* Discovered that the last timestamp of the April file is actually in
May, and assumed this is due to the way SETTLEMENTDATE is the time of
decision about a period that is in the past. The same offset caused the
bug of adding 24h to a date and rolling over to an impossible date
during the calculation of some timestamps. My fix for both of these
issues was to ensure the settlement periods are zero-based and are
identified by their start time, not end time. Now the times displayed
are all 30minutes backward from the source data. This also simplifies
plotting.

* Calculated actual megawatt hour energy difference for the period,
which basically answers the original analytics question that was posed.

* Tested from the command line and found the figures did not show
automatically like they do in Spyder3 IDE, so needed an explicit show()
call on the matplotlib object.